# improve performance of merge