--- a/index.js
+++ b/index.js
@@ -36,17 +36,9 @@ function getKeys(target) {
 	return Object.keys(target).concat(getEnumerableOwnPropertySymbols(target))
 }
 
-function propertyIsOnObject(object, property) {
-	try {
-		return property in object
-	} catch(_) {
-		return false
-	}
-}
-
 // Protects from prototype poisoning and unexpected merging up the prototype chain.
 function propertyIsUnsafe(target, key) {
-	return propertyIsOnObject(target, key) // Properties are safe to merge if they don't exist in the target yet,
+	return key in target // Properties are safe to merge if they don't exist in the target yet,
 		&& !(Object.hasOwnProperty.call(target, key) // unsafe if they exist up the prototype chain,
 			&& Object.propertyIsEnumerable.call(target, key)) // and also unsafe if they're nonenumerable.
 }
@@ -58,12 +50,18 @@ function mergeObject(target, source, options) {
 			destination[key] = cloneUnlessOtherwiseSpecified(target[key], options)
 		})
 	}
+
+	if (typeof target !== 'object' || target === null) {
+		destination = cloneUnlessOtherwiseSpecified(source, options)
+		return destination;
+	}
+
 	getKeys(source).forEach(function(key) {
 		if (propertyIsUnsafe(target, key)) {
 			return
 		}
 
-		if (propertyIsOnObject(target, key) && options.isMergeableObject(source[key])) {
+		if (key in target && options.isMergeableObject(source[key])) {
 			destination[key] = getMergeFunction(key, options)(target[key], source[key], options)
 		} else {
 			destination[key] = cloneUnlessOtherwiseSpecified(source[key], options)